python/finance-processor/pyproject.toml
[project]
name = "finance-processor"
version = "0.1.0"
description = "FastAPI Bank Statement Processor with OCR and OSS integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "oci>=2.0.0",
    "pdf2image>=1.16.0",
    "Pillow>=9.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.build]
include = [
    "app/web/index.html",
]